서치, 뎁스, 스크롤 수정
