+++ src/pages/subPages/faq/faqData.jsx
@@ -67,67 +67,67 @@ const Fcontents = {
     },{
         id : 12,
         fcate: '기타',
-        fqes : 'Q.\u00A0\u00A0[기타] 고객센터 운영 시간',
-        fans : `예매문의 : 티켓링크 고객센터 1588-7890 (10:00~19:00 월~일)\n\n행사문의 : 궁능활용프로그램 콜센터 1522-2295 (10:00~18:00 월~일)`
+        fqes : 'Q.\u00A0\u00A0[기타] 자주하는 질문',
+        fans : `자주하는 질문`
     },
     {
         id : 13,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 주차 장소를 알고 싶어요.',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `경복궁 별빛야행 - 경복궁 동편 유료주차장(주차비 본인부담) 이용가능\n\n창덕궁 달빛기행 - 주차장 따로 없음`
     },
     {
         id : 14,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 축제 당일 비가 오면 어떻게 되나요?',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `우천시 안내\n\n\n행사 당일 오전 11시 기상청 예보 기준\n\n- 1~4mm 예보 시 : 행사 정상진행(정상 진행에 대한 관람객 개별 문자 통지)\n\n- 5mm 이상 예보 시 : 행사 취소(관람객 개별 문자 통지 및 유선 연락)`
     },
     {
         id : 15,
         fcate: '예매',
-        fqes : 'Q.\u00A0\u00A0[예매] 취소 기한을 알고 싶어요.',
+        fqes : 'Q.\u00A0\u00A0[예매] 자주하는 질문',
         fans : `취소 마감시간 : 관람일 전일 오후 5시 이전\n\n\n마감기한 경과 후 취소 및 변경은 불가합니다.\n\n행사의 취소와 환불은 티켓링크 규정을 준수합니다.`
     },
     {
         id : 16,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 참가비는 얼마인가요?',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `경복궁 별빛야행 - 1인 기준 60,000원\n\n창덕궁 달빛기행 - 1인 기준 30,000원`
     },
     {
         id : 17,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 입장권 현장 구매가 가능한가요?',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `100% 사전예약제로 운영되며, 별도의 현장판매는 진행되지 않습니다.`
     },
     {
         id : 18,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 연령제한이 있나요?',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `8세 이상 관람 가능합니다. (취학아동•청소년 보호자 동반)`
     },
     {
         id : 19,
         fcate: '예매',
-        fqes : 'Q.\u00A0\u00A0[예매] 전화예매 대상을 알고 싶어요.',
+        fqes : 'Q.\u00A0\u00A0[예매] 자주하는 질문.',
         fans : `예매대상 : 만 65세 이상, 장애인, 국가유공자\n\n* 티켓링크 : 1588-7890`
     },
     {
         id : 20,
         fcate: '예매',
-        fqes : 'Q.\u00A0\u00A0[예매] 1인당 몇 매까지 예매할 수 있나요?',
+        fqes : 'Q.\u00A0\u00A0[예매] 자주하는 질문',
         fans : `아이디당 최대 2매까지 예매 가능합니다.`
     },
     {
         id : 21,
         fcate: '축제',
-        fqes : 'Q.\u00A0\u00A0[축제] 관람시간이 궁금합니다.',
+        fqes : 'Q.\u00A0\u00A0[축제] 자주하는 질문',
         fans : `경복궁 별빛야행 - 110분\n\n창덕궁 달빛기행 - 100분`
     },
     {
         id : 22,
         fcate: '예매',
-        fqes : 'Q.\u00A0\u00A0[예매] 할인정보가 궁금합니다.',
+        fqes : 'Q.\u00A0\u00A0[예매] 자주하는 질문',
         fans : `장애인 50% 할인(1~3급은 본인 포함 2인, 4급 이하는 본인 한정)\n\n국가유공자 본인 한정 50% 할인\n\n\n현장에서 증빙자료를 확인할 예정이오니 신분증(국가유공자증, 장애인 복지카드 등)을 반드시 지참하여 주시기 바랍니다. 증빙 미지참 시 입장이 불가한 점 양해 부탁드립니다.\n\n꼭 본인이 참여하여야 하며, 참여하지 않으면 현장에서 취소 됩니다.`
     }]
 }
