+++ src/pages/subPages/fesInfo/InfoScroll.jsx
@@ -5,7 +5,7 @@ export default function InfoScroll(props) {
     const { pathname } = useLocation();
 
     useEffect(() => {
-        window.scrollTo(0, 0);
+        window.scrollTo({top:500, left:0, behavior:'smooth'});
     }, [pathname]);
 
     return <>{props.children}</>;
