+++ src/pages/subPages/fesInfo/subFesInfoPages.jsx
@@ -1,20 +1,21 @@
 import { SubFesInfoK, SubFesInfoC, ProgramK, ProgramC, TicketK, TicketC } from "./subFesInfoCom"
-import { useState } from "react";
-import { Link } from 'react-router-dom';
 import SupTop from "../../../components/common/supTop"
 import LineTit from "../../../components/lineTit"
 import {InfoDep, InfoDep2, InfoDep3}  from "./infoDep";
+import InfoScroll from "./InfoScroll";
 
 function Kinfo(){
 
 
     return(
         <>
+        <InfoScroll>
         <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
         <InfoDep one={1}/>
         <InfoDep2 one={1}/>
         <LineTit subtit={'축제정보'}/>
         <SubFesInfoK/>
+        </InfoScroll>
         </>
     )
 }
@@ -23,11 +24,13 @@ function Cinfo(){
 
     return(
         <>
+        <InfoScroll>
         <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
         <InfoDep one={2}/>
         <InfoDep3 one={1}/>
         <LineTit subtit={'축제정보'}/>
         <SubFesInfoC/>
+        </InfoScroll>
         </>
     )
 }
@@ -36,11 +39,13 @@ function Kpro(){
   
   return(
     <>
+    <InfoScroll>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
     <InfoDep one={1}/>
     <InfoDep2 one={2}/>
     <LineTit subtit={'프로그램'}/>
     <ProgramK/>
+    </InfoScroll>
     </>
   )
 }
@@ -49,11 +54,13 @@ function Cpro(){
  
   return(
     <>
+    <InfoScroll>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
     <InfoDep one={2}/>
     <InfoDep3 one={2}/>
     <LineTit subtit={'프로그램'}/>
     <ProgramC/>
+    </InfoScroll>
     </>
   )
 }
@@ -62,11 +69,13 @@ function Kticket(){
  
   return(
     <>
+    <InfoScroll>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
     <InfoDep one={1}/>
     <InfoDep2 one={3}/>
     <LineTit subtit={'예매 안내'}/>
     <TicketK/>
+    </InfoScroll>
     </>
   )
 }
@@ -75,11 +84,13 @@ function Cticket(){
  
   return(
     <>
+    <InfoScroll>
     <SupTop supTopImg={'subFesInfoImage'} supTopTit={'축제소개'}/>
     <InfoDep one={2}/>
     <InfoDep3 one={3}/>
     <LineTit subtit={'예매 안내'}/>
     <TicketC/>
+    </InfoScroll>
     </>
   )
 }
